refactor : 폴더명 변경 redis-repository -> redis
