+++ BE/src/query/query.service.ts
@@ -6,7 +6,7 @@ import { Connection, ResultSetHeader, RowDataPacket } from 'mysql2/promise';
 import { Shell } from '../shell/shell.entity';
 import { UserDBManager } from '../config/query-database/user-db-manager.service';
 import { UsageService } from 'src/usage/usage.service';
-import { ActiveUserRepository } from 'src/redis-repositories/activeUser.repository';
+import { ActiveUserRepository } from 'src/redis/activeUser.repository';
 
 @Injectable()
 export class QueryService {
