+++ BE/src/record/record.module.ts
@@ -5,7 +5,7 @@ import { UserDBModule } from '../config/query-database/user-db.moudle';
 import { UsageModule } from '../usage/usage.module';
 import { TableModule } from 'src/table/table.module';
 import { FileService } from './file.service';
-import { RepositoriesModule } from 'src/redis-repositories/repositories.module';
+import { RepositoriesModule } from 'src/redis/repositories.module';
 
 @Module({
   imports: [RepositoriesModule, UserDBModule, UsageModule, TableModule],
