+++ BE/src/table/table.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TableService } from './table.service';
 import { TableController } from './table.controller';
 import { UserDBModule } from '../config/query-database/user-db.moudle';
-import { RepositoriesModule } from 'src/redis-repositories/repositories.module';
+import { RepositoriesModule } from 'src/redis/repositories.module';
 
 @Module({
   imports: [UserDBModule, RepositoriesModule],
